+++ src/main/java/ua/tour/api/repo/HotelRepository.java
@@ -3,5 +3,5 @@ package ua.tour.api.repo;
 import org.springframework.data.repository.CrudRepository;
 import ua.tour.api.entities.Hotel;
 
-public interface HotelRepository extends CrudRepository<Hotel, Integer> {
+public interface HotelRepository extends CrudRepository<Hotel, Long> {
 }
