+++ src/main/java/ua/tour/api/repo/TourRepository.java
@@ -1,7 +1,11 @@
 package ua.tour.api.repo;
 
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.repository.query.Param;
 import ua.tour.api.entities.Tour;
 
-public interface TourRepository extends CrudRepository<Tour, Integer> {
+import java.util.Date;
+
+public interface TourRepository extends CrudRepository<Tour, Long> {
 }
