+++ src/main/java/ua/tour/api/repo/TourReservationRepository.java
@@ -3,5 +3,5 @@ package ua.tour.api.repo;
 import org.springframework.data.repository.CrudRepository;
 import ua.tour.api.entities.TourReservation;
 
-public interface TourReservationRepository extends CrudRepository<TourReservation, Integer> {
+public interface TourReservationRepository extends CrudRepository<TourReservation, Long> {
 }
